Implement insane difficulty
The score label is bound to the score property of the State class.
If the player is playing on insane difficulty, the canvas will rotate every 5 points.

Class State receives a callback upon instantiation by the Game class.
The callback will be used to alter the game's speed at runtime.